get a clean compile again and pass our first test!

+++ allocate.c
@@ -1,3 +1,4 @@
+#line 1 "allocate.c"
 /**
  * @file allocate.c
  *
@@ -60,7 +61,7 @@ static inline boolean_t should_log() {
  */
 uint8_t* checked_malloc(char* file, int line, uint64_t amount) {
   if (should_log()) {
-    fprintf(stderr, "ALLOCATE %s:%d -- %d\n", file, line, amount);
+    fprintf(stderr, "ALLOCATE %s:%d -- %lu\n", file, line, amount);
   }
   uint8_t* result = malloc(amount);
   if (result == NULL) {
@@ -80,7 +81,8 @@ uint8_t* checked_malloc(char* file, int line, uint64_t amount) {
  */
 void checked_free(char* file, int line, void* pointer) {
   if (should_log()) {
-    fprintf(stderr, "DEALLOCATE %s:%d -- %d\n", file, line, pointer);
+    fprintf(stderr, "DEALLOCATE %s:%d -- %lu\n", file, line,
+            (uint64_t) pointer);
   }
   if (pointer == NULL) {
     fatal_error_impl(file, line, ERROR_MEMORY_FREE_NULL);
